requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
click==8.1.7
fastapi==0.112.1
h11==0.14.0
idna==3.8
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
rsa==4.9
sniffio==1.3.1
starlette==0.38.2
Telethon==1.36.0
typing_extensions==4.12.2
uvicorn==0.30.6
wheel==0.44.0
